requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
gunicorn==21.2.0
Flask-Login==0.6.3
psycopg2-binary
